chore: move validator methods to test-tools from ledger tests
