+++ test-integration/test-ledger-restore/tests/06_delegated_account.rs
@@ -1,4 +1,5 @@
 use cleanass::assert_eq;
+use integration_test_tools::validator::cleanup;
 use std::{path::Path, process::Child};
 
 use integration_test_tools::{expect, tmpdir::resolve_tmp_dir};
@@ -14,7 +15,7 @@ use solana_sdk::{
     signer::Signer,
 };
 use test_ledger_restore::{
-    cleanup, confirm_tx_with_payer_chain, confirm_tx_with_payer_ephem,
+    confirm_tx_with_payer_chain, confirm_tx_with_payer_ephem,
     fetch_counter_chain, fetch_counter_ephem, fetch_counter_owner_chain,
     setup_validator_with_local_remote, wait_for_cloned_accounts_hydration,
     wait_for_ledger_persist, FLEXI_COUNTER_ID, TMP_DIR_LEDGER,
