+++ test-integration/test-ledger-restore/tests/07_commit_delegated_account.rs
@@ -1,4 +1,5 @@
 use cleanass::assert_eq;
+use integration_test_tools::validator::cleanup;
 use std::{path::Path, process::Child};
 
 use integration_test_tools::{expect, tmpdir::resolve_tmp_dir};
@@ -15,7 +16,7 @@ use solana_sdk::{
     signer::Signer,
 };
 use test_ledger_restore::{
-    assert_counter_commits_on_chain, cleanup, confirm_tx_with_payer_chain,
+    assert_counter_commits_on_chain, confirm_tx_with_payer_chain,
     confirm_tx_with_payer_ephem, fetch_counter_chain, fetch_counter_ephem,
     fetch_counter_owner_chain, get_programs_with_flexi_counter,
     setup_validator_with_local_remote, wait_for_cloned_accounts_hydration,
