+++ test-integration/test-ledger-restore/tests/09_restore_different_accounts_multiple_times.rs
@@ -1,4 +1,5 @@
 use cleanass::assert_eq;
+use integration_test_tools::validator::cleanup;
 use std::{path::Path, process::Child};
 
 use integration_test_tools::{expect, tmpdir::resolve_tmp_dir};
@@ -13,7 +14,7 @@ use solana_sdk::{
     native_token::LAMPORTS_PER_SOL, signature::Keypair, signer::Signer,
 };
 use test_ledger_restore::{
-    cleanup, confirm_tx_with_payer_chain, confirm_tx_with_payer_ephem,
+    confirm_tx_with_payer_chain, confirm_tx_with_payer_ephem,
     fetch_counter_chain, fetch_counter_ephem, get_programs_with_flexi_counter,
     setup_validator_with_local_remote, wait_for_cloned_accounts_hydration,
     wait_for_ledger_persist, TMP_DIR_LEDGER,
