+++ test-integration/test-ledger-restore/tests/10_readonly_update_after.rs
@@ -1,7 +1,9 @@
 use cleanass::assert_eq;
 use std::{path::Path, process::Child};
 
-use integration_test_tools::{expect, tmpdir::resolve_tmp_dir};
+use integration_test_tools::{
+    expect, tmpdir::resolve_tmp_dir, validator::cleanup,
+};
 use program_flexi_counter::instruction::{
     create_add_counter_ix, create_add_ix,
 };
@@ -13,7 +15,7 @@ use solana_sdk::{
     native_token::LAMPORTS_PER_SOL, signature::Keypair, signer::Signer,
 };
 use test_ledger_restore::{
-    assert_counter_state, cleanup, confirm_tx_with_payer_chain,
+    assert_counter_state, confirm_tx_with_payer_chain,
     confirm_tx_with_payer_ephem, fetch_counter_chain, fetch_counter_ephem,
     get_programs_with_flexi_counter, setup_validator_with_local_remote,
     wait_for_cloned_accounts_hydration, wait_for_ledger_persist, Counter,
